+++ src/frontend/Player/createPlayer.spec.ts
@@ -1,6 +1,8 @@
-import { Video } from '../types/tracks';
+import { videoMockFactory } from '../utils/tests/factories';
 import { createPlayer } from './createPlayer';
 import { createPlyrPlayer } from './createPlyrPlayer';
+import { createVideojsPlayer } from './createVideojsPlayer';
+import { report } from '../utils/errors/report';
 
 jest.mock('jwt-decode', () => {
   return jest.fn().mockImplementation(() => ({
@@ -16,12 +18,18 @@ jest.mock('../data/appData', () => ({
 }));
 
 jest.mock('./createPlyrPlayer');
+jest.mock('./createVideojsPlayer');
+jest.mock('../utils/errors/report');
 
 describe('createPlayer', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
   it('creates a plyr instance when type player is plyr', () => {
     const ref = 'ref' as any;
     const dispatchPlayerTimeUpdate = jest.fn();
-    const video = {} as Video;
+    const video = videoMockFactory();
 
     createPlayer('plyr', ref, dispatchPlayerTimeUpdate, video);
 
@@ -29,6 +37,32 @@ describe('createPlayer', () => {
       ref,
       dispatchPlayerTimeUpdate,
       video,
+    );
+  });
+
+  it('creates a videojs instance when type player is videojs', () => {
+    const ref = 'ref' as any;
+    const dispatchPlayerTimeUpdate = jest.fn();
+    const video = videoMockFactory();
+
+    createPlayer('videojs', ref, dispatchPlayerTimeUpdate, video);
+
+    expect(createVideojsPlayer).toHaveBeenCalledWith(
+      ref,
+      dispatchPlayerTimeUpdate,
+      video,
+    );
+  });
+
+  it('reports an error if the player is not implemented', () => {
+    const ref = 'ref' as any;
+    const dispatchPlayerTimeUpdate = jest.fn();
+    const video = videoMockFactory();
+
+    createPlayer('unknown', ref, dispatchPlayerTimeUpdate, video);
+
+    expect(report).toHaveBeenCalledWith(
+      Error('player unknown not implemented'),
     );
   });
 });