+++ src/frontend/Player/createPlayer.ts
@@ -1,5 +1,7 @@
 import { VideoPlayerCreator } from '../types/VideoPlayer';
 import { createPlyrPlayer } from './createPlyrPlayer';
+import { createVideojsPlayer } from './createVideojsPlayer';
+import { report } from '../utils/errors/report';
 
 export const createPlayer: VideoPlayerCreator = (
   type,
@@ -10,5 +12,12 @@ export const createPlayer: VideoPlayerCreator = (
   switch (type) {
     case 'plyr':
       return createPlyrPlayer(ref, dispatchPlayerTimeUpdate, video);
+    case 'videojs':
+      const player = createVideojsPlayer(ref, dispatchPlayerTimeUpdate, video);
+      return {
+        destroy: () => player.dispose(),
+      };
+    default:
+      report(new Error(`player ${type} not implemented`));
   }
 };